docs: update README to the project current state
Updated the convenience shortcuts to match the new defined commands in
the makefile.

Added a link to the list of prepopulated accounts.

Added new goals:
- Quick onboard
- Rely on tests
- Continuous delivery

Rewrote the Contribute section, can't expect the community to go around
and checking other links. All the needed information should be on the
repo.

Added a guide on how to setup the development environment.

TODO: write a short guide on how to add new features.