refactor: enforce single call on async methods
Asio executes multiples internal async operations when calling
async_read() or async_write(). If there are more than one operation of
the same type, the execution may interleave causing data corruption. It
doesn't make much sense either since tcp is byte stream oriented, if
there are multiple async reads in place, then would need to resemble the
data from the different buffers in order. Also it's not possible to read
in parallel from a socket, only one read will happen at a time, so the
overhead and complexity adds no benefit.

Adding to that, shall not execute async operations in a block of game
logic, it opens the door to undefined behavior if two async chains are
operating on a shared state. To avoid problems, it's a good practice to
ensure that an event processing is executed completely in a synchronous
manner.

For writes, using a send queue and a single writer loop solves the
problem. The examples suggest to implement an awaitable timer, causing
the send loop to freeze until the send function is called, which cancels
the timer, continuing the send loop.

For reads, there shall be only one read loop, and may be a good thing to
hide the existense of the read operation in the messagestream to ensure
that is not called in a block of game logic. This is because at game
event levels, everything is happening asynchronous, so there is no point
calling read in a game logic block because there is no way to know what
event will come next.

Remove the awaitable decorator from all game logic blocks forcing them
to execute in a synchronous manner.