refactor: use typeid on ievent references
When the typeid function is used on a reference, it returns the
information of the actual object instead of the reference type.

This makes unnecesary to pass the std::type_index between the reactor,
client, serializer and deserializer.

This commit removes that, and that means there is not need of having the
CRTP Event<>, so it also removes this class and it's usage.