build: cmake download dependencies from projects sources
The plan was to store the dependencies on a dedicated folder on the
repository, based on the idea that the external sources some day could
be deleted/lost.

Changed that because it makes difficult to update them. It also happened
that I reformatted those files by accident.

Git submodules could be a good alternative, but that doesn't protect
from the owners deleting the repository.

At the end it's easier to just download from the source.