refactor: promote experimental to main
The experiment proved to be successfull. Replaced the entire src
directory with the content of experimental.

This included deleting all the older code, even tho some of it is still
useful, it will be better to reimplement it again to adapt to the new
codebase.

This change also separate the project structure in three main parts:
- Framework: includes the common base to develop the servers.
- Login: includes the content related to the login server.
- Game: includes the content related to the game server.

Ideally, anything put inside framework should be independent of the game
logic.

With this separation I created three groups of cmake targets, one for
each type of project.

The openao_login_server and openao_game_server are obviously required to
be separated.

I think it's possible to eliminate the openao_framework target
alltogether, and let it be built by who requires it. I am not sure if
makes any difference tho. May change depending on the results.