build: (ab)using docker - build with bind mounts
The application pipeline should be standard. The developers must work in
the same environment used by the CI server. This helps to throw "it
compiles in my machine" out of the table.
The commands used to setup the environment, configure cmake, build and
test should also be shared between the developers and CI.