chore: create a target for the database schema
The database autogenerated (sqlpp11) schema header is shared between the
login and game server.

It doesn't seem proper to include the generation cmake code in both of
them, nor to put only on one, both are dependant but there should be one
generation code snippet.

Was temped to put in framework, but framework shouldn't be tied to
business logic. That's why added a new target for the database schema
only.

If neccesary in the future, could be renamed to 'common' or something
similar and include any other kind of shared component.