chore: cmake fetchcontent using git instead of zip
Changed the way framework pulls its dependencies. It downloading them as
zip files from github. With this change, it downloads them as git
repositories.

The reason is that archives may be deleted and the build will suddenly
fail. The git repository is always there and tags are rarely deleted.

On top of that, it's a more standarized process, some repositories
doesn't publish "Releases" so that would require having two different
ways of pulling dependences, don't like that.