feat: redirect to game server on 'enter game'
When the user press 'Enter game' on the character selection screen, the
server sends a Redirection event with information of the game server
hosting the specific player branch-stage.

This commit adds a command 'EnterGameCommand', adds a event
'RedirectServerEvent'.

Added the handler into accountcontroller.h

The game server information is hardcoded, in the future it should be
retrieved from a service based on the player stage and other properties.

Had to implement the comparator operator for String in order to test the
command.

TODO: Must test the String comparator, should be like a c string
comparer, considering only the values before a null terminator.
Should analyze what would happen then if the String container is used
to store binary values and the comparation function is invoked.