fix: instant disconnection on game server
The client was instantly disconnected after login to the game server.
I don't know the details but I suppose it is because there wasn't a
clear ownership of the unique_ptr. When calling co_await, probably
in the process, the ptr is lost/destroyed.

Should think if there is another way of implementing this, but for
the meanwhile it works (although never destroys the sockets).