build: simplify makefile, add new targets
Added host network to the local docker compose. This allows to connect
the servers to the database regardless if running from the compose stack
or from the host operating system.

Added workdir to the runner image, to copy files inside a meaningful
folder when running test from the makefile.

Improved some targets from the makefile. Added colors. Added new
targets:
- database: launchs the database stack. Useful when developing.
- seed: runs the database seed again. Useful to update with new data.
- login: runs the login server. Useful when working only on game.
- game: runs the game server.