fix: server crashing on player disconnection
Caused when trying to deserialize the incoming buffer.

The root cause of the issue was that BinaryBuffer doesn't do bound
checks, and was trying to read 2 bytes from an empty (size 0) array.
That shall not be allowed. Added a simple check and exception.

The cause of that was Client catches the 'End of file' from
messagestream but because that's inside the read while loop, it then
issues another read, which messagestream returns an empty sized buffer
(because read_buffer_.eof() isn't false and filled with 0s). After that
tries to deserialize that emtpy data, causing the SIGSEGV.

Still to define if should keep processing the already received messages
if the client has disconnected or just drop'em.

TODO:
- Check all the possible SIGSEGV on BinaryBuffer
- Analyze the read function of messagestream for possible bugs.
- Implement custom exceptions (don't use throw c strings)