feat: ensure only one connection per character
There should be only one client binded to a character at any point. This
change makes sure that when a client tries to login, if there is already
a connection binded to that account, then it's closed.

The new client must be closed too. The reason is the disconnection in
asynchronous, once the client disconnection is issued, the client sends
all the pending messages before closing it.

If we issue the disconnection to the old client, and then let the new
client to login, we risk having two clients binded to the same
character for a moment, which could lead to nasty bugs.

To perform a cleanup after a client has disconnected, added a special
event 'ConnectionClosed', which is issued by the Client when the
destructor is called. This is to ensure that no matter what, if the
disconnection was issued by a controller, or because the Client failed
to perform an operation, both go through the same cleanup process.

Also helps that any other operation that disconnects a client is ensured
to go through the same cleanup process.