feat: shared world state, allow player interaction
Added a prototype of what could be a 'interest management'. It's far
from perfect/complete, but at least enables player communication and
it's a starting point to improve.

For the moment it considers the whole server as a single stage, so the
events are broadcasted to anyone connected.

It only allows to spawn other players, doesn't take in consideration
other entities like mobs or sellers.