feat: enable character title change
Accept the event 'set title' from the client, make the local change and
broadcast the update to all other connected players in the stage.

This includes a change in the String class, which didn't had a method
for assigment of types String, caused a problem when the incoming String
was of different size than the current one, because the default
asignation operator replaces everything.
We need to make sure that the assigned buffer maintains it size.