feat: add basic character persistence to database
The characters are now saved into the database when logging out. This
includes if the server is closed gracefully (not sure what would happen
in case of a crash).

For the moment only minimal info is saved, the title, position and
orientation.

Must be updated when adding more features
(stats/equipment/attributes).