feat: spawn npcs
This is a prototype (needs lot of refactoring) to spawn npcs loaded from
the database.

The NPCs are loaded on server start from the database. This probably
should be changed and instead do a lazy load when a stage is
instantiated. It would be also good to remove the entities from memory
when a stage has been idle for X amount of time.

Currently only supports static npcs (no movement, no interaction).